pyproject.toml
[project]
name = "orbit"
version = "0.1.0"
description = ""
authors = [
    {name = "Simon"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv (>=1.2.1,<1.3.0)",
    "requests (>=2.32.4,<2.33.0)",
    "pylette (>=5.1.1,<5.2.0)",
    "climage (>=0.2.2,<0.3.0)",
    "toml (>=0.10.2,<0.11.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
orbit = "orbit.__main__:main"
clean = "rm -r build/ dist/ orbit.spec"